Ajuste no calendário e no calculo do beneficio no orçamento

+++ dashboard/models.py
@@ -89,7 +89,7 @@ class CalendarioMensal(models.Model):
     mes = models.PositiveIntegerField()
     ano = models.PositiveIntegerField()
     funcionario = models.ForeignKey(Colaboradores, on_delete=models.CASCADE)
-    dias_uteis = models.PositiveIntegerField()
+    dias_uteis = models.DecimalField(max_digits=5, decimal_places=2)
     jornada_diaria = models.DecimalField(max_digits=5, decimal_places=2)  # Adicione este campo
     horas_produtivas = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
